+++ src/main/java/com/museu/model/ExpostoEm.java
@@ -14,22 +14,22 @@ public class ExpostoEm {
     @Id
     @NotNull
     @Column(name = "numobj6")
-    private int numObj6;
+    private Long numObj6;
 
     @NotNull
     @Column(name = "nomeexpo")
     private String nomeExpo;
 
-    public ExpostoEm(int numObj6, String nomeExpo) {
+    public ExpostoEm(Long numObj6, String nomeExpo) {
         this.numObj6 = numObj6;
         this.nomeExpo = nomeExpo;
     }
 
-    public int getNumObj6() {
+    public Long getNumObj6() {
         return numObj6;
     }
 
-    public void setNumObj6(int numObj6) {
+    public void setNumObj6(Long numObj6) {
         this.numObj6 = numObj6;
     }
