+++ src/main/java/com/museu/model/ObjetosArte.java
@@ -11,7 +11,7 @@ public class ObjetosArte {
     @Id
     @NotNull
     @Column(name = "numid")
-    private int numId;
+    private Long numId;
 
     @NotNull
     @Column(name = "titulo")
@@ -49,7 +49,7 @@ public class ObjetosArte {
     @Column(name = "nomeart")
     private String nomeArt;
 
-    public ObjetosArte(int numId, String titulo, String descricao, String anoCriacao, String periodoObj, String paisCultura, String estilo, String tipoObjArt, String catObjArt, String nomeArt) {
+    public ObjetosArte(Long numId, String titulo, String descricao, String anoCriacao, String periodoObj, String paisCultura, String estilo, String tipoObjArt, String catObjArt, String nomeArt) {
         this.numId = numId;
         this.titulo = titulo;
         this.descricao = descricao;
@@ -62,11 +62,11 @@ public class ObjetosArte {
         this.nomeArt = nomeArt;
     }
 
-    public int getNumId() {
+    public Long getNumId() {
         return numId;
     }
 
-    public void setNumId(int numId) {
+    public void setNumId(Long numId) {
         this.numId = numId;
     }
